+++ pyccel/codegen/printing/ccode.py
@@ -2391,10 +2391,11 @@ class CCodePrinter(CodePrinter):
             args = args[:1] + self._temporary_args + args[1:]
         else:
             args = self._temporary_args + args
-        self._temporary_args = []
-        args = ', '.join(self._print(ai) for a in args for ai in self.scope.collect_all_tuple_elements(a))
 
-        call_code = f'{func.name}({args})'
+        args_code = ', '.join(self._print(a) for a in args)
+        self._temporary_args = []
+
+        call_code = f'{func.name}({args_code})'
         if func.results.var is not Nil() and \
                 not isinstance(func.results.var.class_type, InhomogeneousTupleType):
             return call_code
@@ -2824,7 +2825,31 @@ class CCodePrinter(CodePrinter):
         return self._print(expr.name)
 
     def _print_FunctionCallArgument(self, expr):
-        return self._print(expr.value)
+        func_call, = expr.get_direct_user_nodes(lambda u: isinstance(u, FunctionCall))
+        idx = func_call.args.index(expr)
+        funcdef_arg = func_call.funcdef.arguments[idx]
+        if isinstance(expr.value, Variable):
+            vals = self.scope.collect_all_tuple_elements(expr.value)
+        elif isinstance(expr.value.class_type, InhomogeneousTupleType):
+            vals = expr.value
+        else:
+            vals = [expr.value]
+        if self.is_c_pointer(funcdef_arg.var):
+            code_elems = []
+            for v in vals:
+                if isinstance(v, Variable):
+                    code_elems.append(self._print(ObjectAddress(v)))
+                elif not self.is_c_pointer(v):
+                    tmp_var = self.scope.get_temporary_variable(v.dtype)
+                    assign = Assign(tmp_var, v)
+                    code_elems = self._print(assign)
+                    self._additional_code += code_elems
+                    code_elems.append(self._print(ObjectAddress(tmp_var)))
+                else:
+                    code_elems.append(self._print(v))
+            return ', '.join(code_elems)
+        else:
+            return ', '.join(self._print(v) for v in vals)
 
     def _print_ObjectAddress(self, expr):
         obj_code = self._print(expr.obj)