+++ pyccel/parser/semantic.py
@@ -2986,11 +2986,15 @@ class SemanticParser(BasicParser):
                 dtype = v.class_type
                 if isinstance(value, Literal) and value is not Nil():
                     value = convert_to_literal(value.python_value, dtype)
-                if isinstance(dtype, InhomogeneousTupleType):
-                    # Raise an error as elements are not yet correctly marked with is_argument.
-                    # This leads to printing errors
-                    errors.report("Inhomogeneous tuples are not yet supported as arguments",
-                            severity='error', symbol=expr)
+
+                tuple_vars = {v}
+                while tuple_vars:
+                    t_var = tuple_vars.pop()
+                    t_var.is_optional = is_optional
+                    t_var.declare_as_argument()
+                    if isinstance(dtype, InhomogeneousTupleType):
+                        tuple_vars.update(self.scope.collect_tuple_element(vi) for vi in t_var)
+
                 if isinstance(dtype, CustomDataType) and not bound_argument:
                     cls = self.scope.find(str(dtype), 'classes')
                     if cls: